Update
Make window movable and resizable on click.

Remove custom event queue. After a while, the event
processing gets delayed more and more, I'm guessing
it has to do with the Low Processor Mode. Reverting
back to CallDeferred and see if it doesn't happen again.
The problem could also be that I accidentally left
a bunch of test programs running in the background,
causing the input delay.